fix: properly switch trigger_teleport base for momentum
The second trigger_teleport is not parsed; parser does not account for multiple entity definitions like this. Solution is to split into two bases with the appropriate base (Trigger or TeleTrigger).

This has the caveat of showing _trigger_teleport/_trigger_teleport_mom in the FGD though. Will be remedied once momentums is properly merged into base engine.