feat: BaseEffectBrush base
Intended for brushes that are used for effects.

In older FGDs this is essentially what brush-based effects ents were including. Some were including parent info and others werent for seemingly no reason, so I just included this in the base for all of them. In the worst case the KV isn't useful.